categories updating in batch

+++ src/main/java/nl/marisabel/backend/categories/repository/CategoryRepository.java
@@ -3,7 +3,7 @@ package nl.marisabel.backend.categories.repository;
 import nl.marisabel.backend.categories.entity.CategoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<CategoryEntity, Integer> {
+public interface CategoryRepository extends JpaRepository<CategoryEntity, Long> {
 
 
 }
